feat(basic): 엔티티 컴포넌트 분리
- `src/basic/components/ProductCard.tsx` 파일에 상품 설명, 할인 정보, 재고 상태 표시 추가
- `src/basic/components/CartItem.tsx` 파일에 할인율 표시 추가
