+++ src/basic/App.tsx
@@ -8,6 +8,8 @@ import { useCart } from './hooks/useCart';
 
 import { useCoupons } from './hooks/useCoupons';
 import { Notification as UINotification } from './components/ui/Notification';
+import { ProductCard } from './components/ProductCard';
+import { CartItem } from './components/CartItem';
 import { Notification, ProductWithUI } from '../types';
 
 const App = () => {
@@ -822,97 +824,15 @@ const App = () => {
                   </div>
                 ) : (
                   <div className="grid grid-cols-1 md:grid-cols-2 lg:grid-cols-3 gap-4">
-                    {filteredProducts.map((product) => {
-                      const remainingStock = getRemainingStock(product);
-
-                      return (
-                        <div
-                          key={product.id}
-                          className="bg-white rounded-lg border border-gray-200 overflow-hidden hover:shadow-lg transition-shadow"
-                        >
-                          {/* 상품 이미지 영역 (placeholder) */}
-                          <div className="relative">
-                            <div className="aspect-square bg-gray-100 flex items-center justify-center">
-                              <svg
-                                className="w-24 h-24 text-gray-300"
-                                fill="none"
-                                stroke="currentColor"
-                                viewBox="0 0 24 24"
-                              >
-                                <path
-                                  strokeLinecap="round"
-                                  strokeLinejoin="round"
-                                  strokeWidth={1}
-                                  d="M4 16l4.586-4.586a2 2 0 012.828 0L16 16m-2-2l1.586-1.586a2 2 0 012.828 0L20 14m-6-6h.01M6 20h12a2 2 0 002-2V6a2 2 0 00-2-2H6a2 2 0 00-2 2v12a2 2 0 002 2z"
-                                />
-                              </svg>
-                            </div>
-                            {product.isRecommended && (
-                              <span className="absolute top-2 right-2 bg-red-500 text-white text-xs px-2 py-1 rounded">
-                                BEST
-                              </span>
-                            )}
-                            {product.discounts.length > 0 && (
-                              <span className="absolute top-2 left-2 bg-orange-500 text-white text-xs px-2 py-1 rounded">
-                                ~
-                                {Math.max(
-                                  ...product.discounts.map((d) => d.rate)
-                                ) * 100}
-                                %
-                              </span>
-                            )}
-                          </div>
-
-                          {/* 상품 정보 */}
-                          <div className="p-4">
-                            <h3 className="font-medium text-gray-900 mb-1">
-                              {product.name}
-                            </h3>
-                            {product.description && (
-                              <p className="text-sm text-gray-500 mb-2 line-clamp-2">
-                                {product.description}
-                              </p>
-                            )}
-
-                            {/* 가격 정보 */}
-                            <div className="mb-3">
-                              <p className="text-lg font-bold text-gray-900">
-                                {formatPrice(product.price, product.id)}
-                              </p>
-                              {product.discounts.length > 0 && (
-                                <p className="text-xs text-gray-500">
-                                  {product.discounts[0].quantity}개 이상 구매시
-                                  할인 {product.discounts[0].rate * 100}%
-                                </p>
-                              )}
-                            </div>
-
-                            {/* 재고 상태 */}
-                            <div className="mb-3">
-                              {remainingStock <= 5 && remainingStock > 0 && (
-                                <p className="text-xs text-red-600 font-medium">
-                                  품절임박! {remainingStock}개 남음
-                                </p>
-                              )}
-                              {remainingStock > 5 && (
-                                <p className="text-xs text-gray-500">
-                                  재고 {remainingStock}개
-                                </p>
-                              )}
-                            </div>
-
-                            {/* 장바구니 버튼 */}
-                            <Button
-                              onClick={() => addToCart(product)}
-                              disabled={remainingStock <= 0}
-                              className="w-full"
-                            >
-                              {remainingStock <= 0 ? '품절' : '장바구니 담기'}
-                            </Button>
-                          </div>
-                        </div>
-                      );
-                    })}
+                    {filteredProducts.map((product) => (
+                      <ProductCard
+                        key={product.id}
+                        product={product}
+                        onAddToCart={addToCart}
+                        getRemainingStock={getRemainingStock}
+                        formatPrice={formatPrice}
+                      />
+                    ))}
                   </div>
                 )}
               </section>
@@ -958,88 +878,15 @@ const App = () => {
                     </div>
                   ) : (
                     <div className="space-y-3">
-                      {cart.map((item) => {
-                        const itemTotal = calculateItemTotal(item, cart);
-                        const originalPrice =
-                          item.product.price * item.quantity;
-                        const hasDiscount = itemTotal < originalPrice;
-                        const discountRate = hasDiscount
-                          ? Math.round((1 - itemTotal / originalPrice) * 100)
-                          : 0;
-
-                        return (
-                          <div
-                            key={item.product.id}
-                            className="border-b pb-3 last:border-b-0"
-                          >
-                            <div className="flex justify-between items-start mb-2">
-                              <h4 className="text-sm font-medium text-gray-900 flex-1">
-                                {item.product.name}
-                              </h4>
-                              <Button
-                                onClick={() => removeFromCart(item.product.id)}
-                                variant="danger"
-                                icon={
-                                  <svg
-                                    className="w-4 h-4"
-                                    fill="none"
-                                    stroke="currentColor"
-                                    viewBox="0 0 24 24"
-                                  >
-                                    <path
-                                      strokeLinecap="round"
-                                      strokeLinejoin="round"
-                                      strokeWidth={2}
-                                      d="M6 18L18 6M6 6l12 12"
-                                    />
-                                  </svg>
-                                }
-                              />
-                            </div>
-                            <div className="flex items-center justify-between">
-                              <div className="flex items-center">
-                                <Button
-                                  onClick={() =>
-                                    updateQuantity(
-                                      item.product.id,
-                                      item.quantity - 1
-                                    )
-                                  }
-                                  variant="secondary"
-                                  className="w-6 h-6"
-                                >
-                                  <span className="text-xs">−</span>
-                                </Button>
-                                <span className="mx-3 text-sm font-medium w-8 text-center">
-                                  {item.quantity}
-                                </span>
-                                <Button
-                                  onClick={() =>
-                                    updateQuantity(
-                                      item.product.id,
-                                      item.quantity + 1
-                                    )
-                                  }
-                                  variant="secondary"
-                                  className="w-6 h-6"
-                                >
-                                  <span className="text-xs">+</span>
-                                </Button>
-                              </div>
-                              <div className="text-right">
-                                {hasDiscount && (
-                                  <span className="text-xs text-red-500 font-medium block">
-                                    -{discountRate}%
-                                  </span>
-                                )}
-                                <p className="text-sm font-medium text-gray-900">
-                                  {Math.round(itemTotal).toLocaleString()}원
-                                </p>
-                              </div>
-                            </div>
-                          </div>
-                        );
-                      })}
+                      {cart.map((item) => (
+                        <CartItem
+                          key={item.product.id}
+                          item={item}
+                          onUpdateQuantity={updateQuantity}
+                          onRemove={removeFromCart}
+                          calculateItemTotal={(cartItem) => calculateItemTotal(cartItem, cart)}
+                        />
+                      ))}
                     </div>
                   )}
                 </section>
